+++ src/queries/list_workers.rs
@@ -1,5 +1,5 @@
 use apalis_core::backend::{BackendExt, ListWorkers, RunningWorker};
-use chrono::{DateTime, Utc};
+use apalis_sql::{DateTime, DateTimeExt};
 use futures::TryFutureExt;
 use ulid::Ulid;
 
@@ -9,8 +9,8 @@ pub struct WorkerRow {
     pub worker_type: String,
     pub storage_name: String,
     pub layers: Option<String>,
-    pub last_seen: DateTime<Utc>,
-    pub started_at: Option<DateTime<Utc>>,
+    pub last_seen: DateTime,
+    pub started_at: Option<DateTime>,
 }
 
 use crate::{CompactType, PgContext, PostgresStorage};
@@ -42,8 +42,11 @@ where
                     .map(|w| RunningWorker {
                         id: w.id,
                         backend: w.storage_name,
-                        started_at: w.started_at.map(|t| t.timestamp()).unwrap_or_default() as u64,
-                        last_heartbeat: w.last_seen.timestamp() as u64,
+                        started_at: w
+                            .started_at
+                            .map(|t| t.to_unix_timestamp())
+                            .unwrap_or_default() as u64,
+                        last_heartbeat: w.last_seen.to_unix_timestamp() as u64,
                         layers: w.layers.unwrap_or_default(),
                         queue: w.worker_type,
                     })
@@ -73,8 +76,11 @@ where
                     .map(|w| RunningWorker {
                         id: w.id,
                         backend: w.storage_name,
-                        started_at: w.started_at.map(|t| t.timestamp()).unwrap_or_default() as u64,
-                        last_heartbeat: w.last_seen.timestamp() as u64,
+                        started_at: w
+                            .started_at
+                            .map(|t| t.to_unix_timestamp())
+                            .unwrap_or_default() as u64,
+                        last_heartbeat: w.last_seen.to_unix_timestamp() as u64,
                         layers: w.layers.unwrap_or_default(),
                         queue: w.worker_type,
                     })